Fixed an issue
Whenever a user reloads the page, performQuery in the App component will not receive any arguments, so the query string in the fetch call won't have any tags, meaning random irrelevant photos will be returned. I solved it by making an if statement: if no arguments are passed to performQuery, then get the path from the URL without the /, and pass that as an argument to performQuery on every page reload.